requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
psycopg2-binary==2.9.9